+++ zee5app-springboot/src/main/java/com/zee/zee5app/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.zee.zee5app.repository;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -15,6 +15,6 @@ public interface UserRepository extends JpaRepository<Register, String> {
 //	Boolean existsByContactNumber(BigDecimal contactNumber);
 	//checking both condition using AND
 	// this can also be done for other conditions like OR
-	Boolean existsByEmailAndContactNumber(String email,BigDecimal contactNumber);
+	Boolean existsByEmailAndContactNumber(String email,BigInteger contactNumber);
 	
 }
